+++ src/main/java/baseball/domain/Referee.java
@@ -1,10 +1,9 @@
-package baseball.domain.referee;
+package baseball.domain;
 
 import static baseball.domain.numbers.NumbersConstant.MAX_INDEX;
 import static baseball.domain.numbers.NumbersConstant.MIN_INDEX;
 
 import baseball.domain.numbers.Numbers;
-import baseball.domain.result.Result;
 import java.util.List;
 import java.util.stream.IntStream;
